Compute the device factory and store it in the database at import time
Computing the device factory on the fly requires parsing and typechecking
ThingTalk code, which we'd rather avoid if we can. Instead, we can
compute the factory once and store it in the database.

Fixes #18
