+++ model/device.js
@@ -124,17 +124,17 @@ module.exports = {
 
     getByCategoryWithCode(client, category, org) {
         if (org !== null && org.is_admin) {
-            return db.selectAll(client, "select d.primary_kind, d.category, d.name, dcv.code from device_class d, "
+            return db.selectAll(client, "select d.primary_kind, d.category, d.name, dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv where d.id = dcv.device_id and category = ? "
                 + "and d.developer_version = dcv.version order by name", [category]);
         } else if (org !== null) {
-            return db.selectAll(client, "select d.primary_kind, d.category, d.name, dcv.code from device_class d, "
+            return db.selectAll(client, "select d.primary_kind, d.category, d.name, dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv where d.id = dcv.device_id and "
                 + "((dcv.version = d.developer_version and d.owner = ?) or "
                 + " (dcv.version = d.approved_version and d.owner <> ?)) "
                 + "and category = ? order by name", [org.id, org.id, category]);
         } else {
-            return db.selectAll(client, "select d.primary_kind, d.category, d.name, dcv.code from device_class d, "
+            return db.selectAll(client, "select d.primary_kind, d.category, d.name, dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv where d.id = dcv.device_id and "
                 + "dcv.version = d.approved_version and category = ? order by name", [category]);
         }
@@ -142,17 +142,17 @@ module.exports = {
 
     getBySubcategoryWithCode(client, category, org) {
         if (org !== null && org.is_admin) {
-            return db.selectAll(client, "select d.primary_kind, d.category, d.name, dcv.code from device_class d, "
+            return db.selectAll(client, "select d.primary_kind, d.category, d.name, dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv where d.id = dcv.device_id and subcategory = ? "
                 + "and d.developer_version = dcv.version order by name", [category]);
         } else if (org !== null) {
-            return db.selectAll(client, "select d.primary_kind, d.category, d.name, dcv.code from device_class d, "
+            return db.selectAll(client, "select d.primary_kind, d.category, d.name, dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv where d.id = dcv.device_id and "
                 + "((dcv.version = d.developer_version and d.owner = ?) or "
                 + " (dcv.version = d.approved_version and d.owner <> ?)) "
                 + "and subcategory = ? order by name", [org.id, org.id, category]);
         } else {
-            return db.selectAll(client, "select d.primary_kind, d.category, d.name, dcv.code from device_class d, "
+            return db.selectAll(client, "select d.primary_kind, d.category, d.name, dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv where d.id = dcv.device_id and "
                 + "dcv.version = d.approved_version and subcategory = ? order by name", [category]);
         }
@@ -160,7 +160,7 @@ module.exports = {
 
     getAllApprovedWithCode(client, org, start, end) {
         if (org !== null && org.is_admin) {
-            const query = "select d.primary_kind, d.category, d.name, dcv.code from device_class d, "
+            const query = "select d.primary_kind, d.category, d.name, dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv where d.id = dcv.device_id and "
                 + "dcv.version = d.developer_version order by d.name";
             if (start !== undefined && end !== undefined) {
@@ -170,7 +170,7 @@ module.exports = {
                 return db.selectAll(client, query, []);
             }
         } else if (org !== null) {
-            const query = "select d.primary_kind, d.category, d.name, dcv.code from device_class d, "
+            const query = "select d.primary_kind, d.category, d.name, dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv where d.id = dcv.device_id and "
                 + "((dcv.version = d.developer_version and d.owner = ?) or "
                 + " (dcv.version = d.approved_version and d.owner <> ?)) order by d.name";
@@ -181,7 +181,7 @@ module.exports = {
                 return db.selectAll(client, query, [org.id, org.id]);
             }
         } else {
-            const query = "select d.primary_kind, d.category, d.name, dcv.code from device_class d, "
+            const query = "select d.primary_kind, d.category, d.name, dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv where d.id = dcv.device_id and "
                 + "dcv.version = d.approved_version order by d.name";
             if (start !== undefined && end !== undefined) {
@@ -354,13 +354,13 @@ module.exports = {
 
     getDevicesForSetup(client, names, org) {
         if (org !== null && org.is_admin) {
-            const query = "(select d.*, d.primary_kind as for_kind,dcv.code from device_class d, "
+            const query = "(select d.*, d.primary_kind as for_kind,dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv where d.id = dcv.device_id and "
                 + "dcv.version = d.developer_version and "
-                + "d.primary_kind in (?)) union all (select d.*, dck.kind as for_kind,dcv.code from device_class d, "
+                + "d.primary_kind in (?)) union all (select d.*, dck.kind as for_kind,dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv, device_class_kind dck where dck.device_id = d.id and d.id = dcv.device_id and "
                 + "dcv.version = d.developer_version and "
-                + "dck.kind in (?)) union all (select d.*, dck2.kind as for_kind,dcv.code from device_class d, "
+                + "dck.kind in (?)) union all (select d.*, dck2.kind as for_kind,dcv.code, dcv.factory from device_class d, "
                 + "device_class d2, device_class_kind dck2, "
                 + "device_code_version dcv, device_class_kind dck where dck.device_id = d.id and d.id = dcv.device_id and "
                 + "dcv.version = d.developer_version and "
@@ -368,15 +368,15 @@ module.exports = {
                 + "dck2.kind in (?))";
             return db.selectAll(client, query, [names, names, names]);
         } else if (org !== null) {
-            const query = "(select d.*, d.primary_kind as for_kind,dcv.code from device_class d, "
+            const query = "(select d.*, d.primary_kind as for_kind,dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv where d.id = dcv.device_id and "
                 + "((dcv.version = d.developer_version and d.owner = ?) or "
                 + " (dcv.version = d.approved_version and d.owner <> ?)) and "
-                + "d.primary_kind in (?)) union all (select d.*, dck.kind as for_kind,dcv.code from device_class d, "
+                + "d.primary_kind in (?)) union all (select d.*, dck.kind as for_kind,dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv, device_class_kind dck where dck.device_id = d.id and d.id = dcv.device_id and "
                 + "((dcv.version = d.developer_version and d.owner = ?) or "
                 + " (dcv.version = d.approved_version and d.owner <> ?)) and "
-                + "dck.kind in (?)) union all (select d.*, dck2.kind as for_kind,dcv.code from device_class d, "
+                + "dck.kind in (?)) union all (select d.*, dck2.kind as for_kind,dcv.code, dcv.factory from device_class d, "
                 + "device_class d2, device_class_kind dck2, "
                 + "device_code_version dcv, device_class_kind dck where dck.device_id = d.id and d.id = dcv.device_id and "
                 + "((dcv.version = d.developer_version and d.owner = ?) or "
@@ -385,13 +385,13 @@ module.exports = {
                 + "dck2.kind in (?))";
             return db.selectAll(client, query, [org.id, org.id, names, org.id, org.id, names, org.id, org.id, names]);
         } else {
-            const query = "(select d.*, d.primary_kind as for_kind,dcv.code from device_class d, "
+            const query = "(select d.*, d.primary_kind as for_kind,dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv where d.id = dcv.device_id and "
                 + "dcv.version = d.approved_version and "
-                + "d.primary_kind in (?)) union all (select d.*, dck.kind as for_kind,dcv.code from device_class d, "
+                + "d.primary_kind in (?)) union all (select d.*, dck.kind as for_kind,dcv.code, dcv.factory from device_class d, "
                 + "device_code_version dcv, device_class_kind dck where dck.device_id = d.id and d.id = dcv.device_id and "
                 + "dcv.version = d.approved_version and "
-                + "dck.kind in (?)) union all (select d.*, dck2.kind as for_kind,dcv.code from device_class d, "
+                + "dck.kind in (?)) union all (select d.*, dck2.kind as for_kind,dcv.code, dcv.factory from device_class d, "
                 + "device_class d2, device_class_kind dck2, "
                 + "device_code_version dcv, device_class_kind dck where dck.device_id = d.id and d.id = dcv.device_id and "
                 + "dcv.version = d.approved_version and "
